getJSComponent fails if called immediately after createChildPrefab #991
